Move close button color logic to router.js
The code that sets the color of the close button in a banner, based on the banner's background color, has been moved from banner.js to router.js. The logic remains the same, but the change improves code organization by centralizing the manipulation of UI elements in the router file.