Merging for 03-28 release (#198)
* Move close button color logic to router.js

The code that sets the color of the close button in a banner, based on the banner's background color, has been moved from banner.js to router.js. The logic remains the same, but the change improves code organization by centralizing the manipulation of UI elements in the router file.

* [ALS-5580] Fix banner v2

- Move some $ around
- Make x contrast

* Remove virtual host configuration files (#182)

Virtual host configuration files have been removed from all relevant Dockerfiles for the PIC-SURE platform. This step simplifies the configuration by eliminating files that were previously copied into the Docker images. The necessary directives are instead being added directly in the Dockerfiles.

* [ALS-5902] Update logout redirect URL in header.js (#190)

* [ALS-5902] Update logout redirect URL in header.js

This update modifies the 'gotoLogin' function in header.js to use the logoutRedirect URL from settings, if it exists. If it doesn't exist, it will default to the "/psamaui/login" location. This change provides flexibility in setting custom logout redirect URLs.

* Update pic-sure-hpds-ui/src/main/webapp/picsureui/header/header.js

* [ALS-5332] Add Manual Study Button (#191)

* Add userDetails.js override (#192)

* [ALS-5974] Variant search fix for multiple variants

* [ALS-6073] Display variant filters in Data Requests

---------

Co-authored-by: Luke Sikina <lucas.skina@gmail.com>
Co-authored-by: James <Jamestp19@gmail.com>